Further validation improvements

+++ src/main/java/pl/lodz/pas/librarianrest/web/controllers/LendingsController.java
@@ -1,5 +1,6 @@
 package pl.lodz.pas.librarianrest.web.controllers;
 
+import pl.lodz.pas.librarianrest.Utils;
 import pl.lodz.pas.librarianrest.web.controllers.objects.MultipleOperationsResult;
 import pl.lodz.pas.librarianrest.services.LendingsService;
 import pl.lodz.pas.librarianrest.services.dto.LendEventDto;
@@ -8,7 +9,6 @@ import javax.ejb.Stateless;
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
 import javax.ws.rs.*;
 import javax.ws.rs.core.Response;
 import java.util.Collections;
@@ -33,7 +33,10 @@ public class LendingsController {
     @Path("lending/return")
     public Response returnLendings(@NotNull List<String> ids) {
 
-        if (ids.contains(null)) return Response.status(Response.Status.BAD_REQUEST).build();
+        if (ids.contains(null) || !ids.stream().allMatch(Utils::isValidUuid)) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
+
 
         var count = service.returnLendings(ids);
 
@@ -43,6 +46,10 @@ public class LendingsController {
     @DELETE
     @Path("lending/{id}")
     public Response deleteNotReturnedLendings(@NotNull @PathParam("id") String id) {
+
+        if (!Utils.isValidUuid(id)) {
+            return Response.status(Response.Status.BAD_REQUEST).build();
+        }
 
         var i = service.removeNotReturnedLendings(Collections.singletonList(id));
